Move React components from 'frontend/src/App.jsx' to their own files in the 'frontend/src/components' directory
